backend/requirements.txt
# ============================================
# Backend Python - ML/NLP Dependencies
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# PostgreSQL Vector Extension
pgvector==0.2.4

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Machine Learning & NLP
sentence-transformers==2.3.1
scikit-learn==1.4.0
scipy==1.12.0

# NLP Libraries
spacy==3.7.2
# Pour télécharger le modèle français: python -m spacy download fr_core_news_lg

# Vector Search
faiss-cpu==1.7.4

# Text Processing
nltk==3.8.1
unidecode==1.3.8

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-magic==0.4.27  # File type detection

# AI/LLM APIs (pour proxy backend sécurisé)
anthropic==0.18.1  # Claude API
google-generativeai==0.3.2  # Gemini API

# Logging & Monitoring
loguru==0.7.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

